% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqtrial_matchit.R
\name{alltreated}
\alias{alltreated}
\title{Get all times at which people initiate treatment}
\usage{
alltreated(id_sym, time_sym, treated_sym, data)
}
\arguments{
\item{id_sym}{A variable given as a symbol}

\item{time_sym}{A variable given as a symbol}

\item{treated_sym}{A variable given as a symbol}

\item{data}{A data frame in which these variables exist. All variables must be in this data frame.}
}
\value{
A data frame
}
\description{
Get all times at which people initiate treatment
}
\examples{
alltreated(
id_var ="id",
time_var = "time",
treated_var = "treated",
data = dummydata
)
}
