% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_splitter.R
\name{data_splitter}
\alias{data_splitter}
\title{Split a Survival Data Set}
\usage{
data_splitter(
  data,
  id = id,
  fup = fup,
  status = status,
  treatment,
  tt_treatment
)
}
\arguments{
\item{data}{a data frame}

\item{id}{column with unique patient identifier (defaults to id)}

\item{fup}{column with time of follow-up (defaults to fup)}

\item{status}{column with the status indicator (defaults to status)}

\item{treatment}{column with the treatment indicator}

\item{tt_treatment}{column with time to treatment}
}
\value{
\code{data_splitter} returns the splitted data in a tibble
}
\description{
This function reshapes data to a long format by splitting each record into
multiple subrecords at intervals of length 1. The resulting dataset will have
a new column 'tstart' and 'tend' with event status and treatment status for
each interval
}
\examples{
data(paxlovid)
data_splitter(data = paxlovid,
              treatment = A,
              tt_treatment = tt_A)

data <- data_simulator(n = 500, n_visit = 5)
data_splitter(data = data,
              treatment = A,
              tt_treatment = tt_A)
}
\author{
Linda Nab, \email{lindanab4@gmail.com}
}
