% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postmatch.R
\name{matchsuccess}
\alias{matchsuccess}
\title{Calculate number of treated people who were matched per trial}
\usage{
matchsuccess(seqtrial_obj)
}
\arguments{
\item{seqtrial_obj}{An object with class seqtrial}
}
\value{
A data frame
}
\description{
Calculate number of treated people who were matched per trial
}
\examples{

seqtrial_example <- seqtrial_matchit(
id_var ="id",
time_var = "time",
treated_var = "treated",
matching_vars = c("age_grp", "biomarker", "sex"),
data = dummydata
)
matchsuccess(seqtrial_example)
}
