% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{paxlovid}
\alias{paxlovid}
\title{Paxlovid data}
\format{
A data frame with 2 rows and 6 variables:
\describe{
  \item{id}{Unique patient identifier}
  \item{fup}{Days of follow-up}
  \item{status}{Event status indicator}
  \item{A}{Treatment indicator}
  \item{tt_A}{Time to treatment (day at which treatment is initiated)}
  \item{L}{Sex}
}
}
\usage{
paxlovid
}
\description{
Follow-up time, event status indicator, treatment status indicator, time to
treatment and sex of individuals in a study investigating the effectiveness
of paxlovid use versus no-use.
}
\details{
This is a dummy data set
}
\examples{
data("paxlovid", package = "seqtrial")
}
\keyword{datasets}
