% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqtrial-package.R
\docType{package}
\name{seqtrial-package}
\alias{seqtrial}
\alias{seqtrial-package}
\title{seqtrial: Implementing the Sequential Trial Approach for Observational Studies}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Elsie Horne \email{elsiehorne@gmail.com} (\href{https://orcid.org/0000-0002-5545-7628}{ORCID})

}
\keyword{internal}
