% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqtrial_matchit.R
\name{seqtrial_matchit}
\alias{seqtrial_matchit}
\title{Perform matching according to the sequential trial approach for target trial emulation}
\usage{
seqtrial_matchit(data)
}
\arguments{
\item{data}{A data frame with class "seqtrial_fmt".}
}
\value{
A data frame with class "seqtrial_matched_fmt"
}
\description{
Perform matching according to the sequential trial approach for target trial emulation
}
\examples{
data("dummydata")

# for treated vs untreated comparison
data_seqtrial_tu <- seqtrial_formatter(
id_var ="id",
time_var = "time",
treated_var = "treated",
matching_vars = c("age_grp", "biomarker", "sex"),
other_vars = NULL,
data = dummydata
)
data_seqtrial_tu_matched <- seqtrial_matchit(data_seqtrial_tu)

# for treatment A vs treatment B comparison
data_seqtrial_ab <- seqtrial_formatter(
id_var ="id",
time_var = "time",
treated_var = "treated",
treatment_var = "treatment",
treatment_vals = c("A", "B"),
matching_vars = c("age_grp", "biomarker", "sex"),
other_vars = NULL,
data = dummydata
)
data_seqtrial_ab_matched <- seqtrial_matchit(data_seqtrial_ab)
}
