% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqtrial_matchit.R
\name{summary.seqtrial_matchit}
\alias{summary.seqtrial_matchit}
\title{View a balance summary of a \code{seqtrial_matchit} object for each trial}
\usage{
\method{summary}{seqtrial_matchit}(seqtrial_matchit_obj, ...)
}
\arguments{
\item{seqtrial_matchit_obj}{a \code{seqtrial_matchit} object; the output of a call to \code{\link[=seqtrial_matchit]{seqtrial_matchit()}}}

\item{\dots}{additional arguments passed to \code{\link{summary.matchit}}}
}
\value{
For \code{seqtrial_matchit} objects, a list with the following components:

\item{call}{the original call to \code{\link[=matchit]{matchit()}}}
\item{nn}{a tibble of the sample sizes in the original (unmatched) and
matched samples per trial}
\item{sum.all}{if \code{un = TRUE}, a tibble of balance statistics for each
covariate in the original (unmatched) sample per trial}
\item{sum.matched}{a tibble of balance statistics for each covariate in the
matched sample per trial}
\item{reduction}{if \code{improvement = TRUE}, a tibble of the percent
reduction in imbalance for each covariate in the matched sample}
}
\description{
View a balance summary of a \code{seqtrial_matchit} object for each trial
}
