% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_constructor.R
\name{trial_constructor}
\alias{trial_constructor}
\title{Construct a Sequence of Trials}
\usage{
trial_constructor(
  data_splitted,
  n,
  id = id,
  status = status,
  treatment,
  censor = FALSE
)
}
\arguments{
\item{data_splitted}{a data frame that has been split using data_splitter()}

\item{n}{number of trials to construct, starting at baseline, n => 1}

\item{id}{column with unique patient identifier (defaults to id)}

\item{status}{column with the status indicator (defaults to status)}

\item{treatment}{column with the treatment indicator}

\item{censor}{boolean indicating if individuals are censored when they
deviate from their baseline treatment}
}
\value{
\code{trial_constructor} returns a data frame in which the sequence
of trials has been contructed
}
\description{
Construct a sequence of trials
}
\examples{
# split data
data(paxlovid)
paxlovid_splitted <-
  data_splitter(data = paxlovid,
                treatment = A,
                tt_treatment = tt_A)
# construct trials
trial_constructor(data = paxlovid_splitted,
                  n = 5,
                  treatment = A)

# same using simulated data
data <- data_simulator(n = 500, n_visit = 5)
data_splitted <-
  data_splitter(data = data,
                treatment = A,
                tt_treatment = tt_A)
data_splitted <-
  data_splitted \%>\%
  dplyr::mutate(
    L = dplyr::case_when(tstart == 0 ~ L0,
                         tstart == 1 ~ L1,
                         tstart == 2 ~ L2,
                         tstart == 3 ~ L3,
                         tstart == 4 ~ L4)) \%>\%
  dplyr::select(-dplyr::starts_with("L"))
trial_constructor(data = data_splitted,
                  n = 5,
                  treatment = A)
}
\author{
Linda Nab, \email{lindanab4@gmail.com}
}
