% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqtrial_matchit.R
\name{seqtrial_formatter}
\alias{seqtrial_formatter}
\title{Formatting a dataset for the sequential trials approach}
\usage{
seqtrial_formatter(
  id_var,
  time_var,
  treated_var,
  treatment_var = NULL,
  treatment_vals = NULL,
  outcome_vars = NULL,
  censor_vars = NULL,
  keep_vars,
  data
)
}
\arguments{
\item{id_var}{The id variable given as a character vector.}

\item{time_var}{The time variable given as a character vector. Must be a numeric variable.}

\item{treated_var}{The variable indicating the time of treatment initiation. Must be coercible to 0 / 1. Takes the value 0 until the time of treatment initiation, and 1 on and after the time of treatment initiation.}

\item{treatment_var}{The treatment variable given as a character vector. Must be NA before treatment initiated. Must be NULL if comparison is treated / untreated.}

\item{treatment_vals}{Values to define the treatment groups in a "treatment A vs treatment B" comparison. Must match the nonmissing values in treated_var. Must be NULL if treatment_var is NULL.}

\item{outcome_vars}{Outcome variables given as a character vector. Must specify at least one outcome variable. Variables must be coercible to 0 / 1, taking the value 0 before the outcome event and 1 at the time of the event. Only the first occurrence of the event will be used, so it does not matter whether the value returns to 0 after the event, or remains at 1.}

\item{censor_vars}{Censoring variables given as a character vector. Variables must be coercible to 0 / 1, taking the value 0 before the censoring event and 1 at the time of the event. Only the first occurrence of the event will be used, so it does not matter whether the value returns to 0 after the event, or remains at 1.}

\item{keep_vars}{The variables to keep in the dataset given as a character vector. Any variables used in matching, assessing balance post-matching or used as covariates in the model should be specified here. At least one variable must be specified.}

\item{data}{A data frame in which these variables exist. All variables must be in this data frame.}
}
\value{
A data frame with class "seqtrial_fmt"
}
\description{
Formatting a dataset for the sequential trials approach
}
\examples{
data("dummydata")

# for treated vs untreated comparison
data_seqtrial_tu <- seqtrial_formatter(
id_var ="id",
time_var = "time",
treated_var = "treated",
outcome_vars = death,
censor_vars = NULL,
keep_vars = c("age_grp", "biomarker_grp", "sex"),
data = dummydata
)

# for treatment A vs treatment B comparison
data_seqtrial_ab <- seqtrial_formatter(
id_var ="id",
time_var = "time",
treated_var = "treated",
treatment_var = "treatment",
treatment_vals = c("A", "B"),
keep_vars = c("age_grp", "biomarker_grp", "sex"),
data = dummydata
)
}
