% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqtrial_matchit.R
\name{seqtrial_matchit}
\alias{seqtrial_matchit}
\title{Perform matching according to the sequential trial approach for target trial emulation}
\usage{
seqtrial_matchit(id_var, time_var, treated_var, matching_vars, data)
}
\arguments{
\item{id_var}{A variable given as a character vector.}

\item{time_var}{A variable given as a character vector.}

\item{treated_var}{A variable given as a character vector.}

\item{matching_vars}{A variable given as a character vector.}

\item{data}{A data frame in which these variables exist. All variables must be in this data frame.}
}
\value{
A list
}
\description{
Perform matching according to the sequential trial approach for target trial emulation
}
\examples{
seqtrial_example <- seqtrial_matchit(
id_var ="id",
time_var = "time",
treated_var = "treated",
matching_vars = c("age_grp", "biomarker", "sex"),
data = dummydata
)
}
